+++ tests/api/query.py
@@ -13,7 +13,7 @@ class TestApi(unittest.TestCase):
                              'https://api-us.logtrust.com/search/query')
         self.key = os.getenv('DEVO_API_KEY', None)
         self.secret = os.getenv('DEVO_API_SECRET', None)
-        self.token = os.getenv('DEVO_API_TOKEN', None)
+        self.token = os.getenv('DEVO_AUTH_TOKEN', None)
         self.query_id = os.getenv('DEVO_API_QUERYID', None)
 
     def test_from_config(self):
@@ -22,37 +22,44 @@ class TestApi(unittest.TestCase):
         api.close()
 
     def test_query(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
+        result = api.query(query=self.query, stream=False, response="json")
+        self.assertIsNotNone(result)
+        self.assertTrue(len(json.loads(result)['object']) > 0)
+        api.close()
+
+    def test_token(self):
+        api = Client(token=self.token, url=self.uri)
         result = api.query(query=self.query, stream=False, response="json")
         self.assertIsNotNone(result)
         self.assertTrue(len(json.loads(result)['object']) > 0)
         api.close()
 
     def test_query_id(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
         result = api.query(query_id=self.query_id,
                               stream=False, response="json")
         self.assertIsNotNone(result)
-        self.assertNotEquals(result, {})
+        self.assertNotEqual(result, {})
         self.assertEqual(type(len(json.loads(result)['object'])), type(1))
         api.close()
 
     def test_query_yesterday_to_today(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
         result = api.query(query=self.query, dates={'from': 'yesterday()', 'to':'today()'},
                            stream=False, response="json")
         self.assertIsNotNone(result)
         self.assertEqual(len(json.loads(result)['object']), 1)
 
     def test_query_from_seven_days(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
         result = api.query(query=self.query, dates={'from': 'now()-7*day()', 'to':'now()'},
                            stream=False, response="json")
         self.assertIsNotNone(result)
         self.assertEqual(len(json.loads(result)['object']), 1)
 
     def test_query_from_fixed_dates(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
         result = api.query(query=self.query,
                            dates={'from': strftime("%Y-%m-%d", gmtime()),
                                  'to': strftime("%Y-%m-%d %H:%M:%S", gmtime())},
@@ -61,7 +68,7 @@ class TestApi(unittest.TestCase):
         self.assertEqual(len(json.loads(result)['object']), 1)
 
     def test_stream_query(self):
-        api = Client(self.key, self.secret, self.uri)
+        api = Client(key=self.key, secret=self.secret, url=self.uri)
         result = api.query(query=self.query, response="json/simple")
         self.assertTrue(isinstance(result, Buffer))
         api.close()